Add common module, fix namespace
