+++ .gitignore
@@ -14,6 +14,7 @@
 /.idea/kotlinc.xml
 /.idea/other.xml
 /.idea/gradle.xml
+/.idea/migrations.xml
 .DS_Store
 /build
 /captures
