+++ settings.gradle.kts
@@ -19,4 +19,5 @@ rootProject.name = "TextFlow"
 include(":demo")
 include(":textflow")
 include(":textflow-material3")
+include(":textflow-common")
