+++ textflow-common/build.gradle.kts
@@ -43,9 +43,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":textflow-common"))
-
     implementation(libs.compose.ui)
     implementation(libs.compose.foundation)
-    implementation(libs.compose.material)
 }
