+++ textflow-common/src/main/kotlin/eu/wewox/textflow/TextFlowLayoutObstacleAlignment.kt
@@ -1,0 +1,16 @@
+package eu.wewox.textflow
+
+/**
+ * The allowed alignment for an "obstacle" inside the [TextFlowLayout] composable.
+ */
+public enum class TextFlowLayoutObstacleAlignment {
+    /**
+     * Obstacle is aligned in the top start corner.
+     */
+    TopStart,
+
+    /**
+     * Obstacle is aligned in the top end corner.
+     */
+    TopEnd,
+}
