+++ textflow-material3/build.gradle.kts
@@ -43,6 +43,8 @@ android {
 }
 
 dependencies {
+    implementation(project(":textflow-common"))
+
     implementation(libs.compose.ui)
     implementation(libs.compose.foundation)
     implementation(libs.compose.material3)
